Fix boost split after rebase
