+++ src/stats/calculators/boost_stats.rs
@@ -1,0 +1,75 @@
+use super::*;
+
+impl BoostStats {
+    pub fn average_boost_amount(&self) -> f32 {
+        if self.tracked_time == 0.0 {
+            0.0
+        } else {
+            self.boost_integral / self.tracked_time
+        }
+    }
+
+    pub fn bpm(&self) -> f32 {
+        if self.tracked_time == 0.0 {
+            0.0
+        } else {
+            self.amount_collected * 60.0 / self.tracked_time
+        }
+    }
+
+    fn pct(&self, value: f32) -> f32 {
+        if self.tracked_time == 0.0 {
+            0.0
+        } else {
+            value * 100.0 / self.tracked_time
+        }
+    }
+
+    pub fn zero_boost_pct(&self) -> f32 {
+        self.pct(self.time_zero_boost)
+    }
+
+    pub fn hundred_boost_pct(&self) -> f32 {
+        self.pct(self.time_hundred_boost)
+    }
+
+    pub fn boost_0_25_pct(&self) -> f32 {
+        self.pct(self.time_boost_0_25)
+    }
+
+    pub fn boost_25_50_pct(&self) -> f32 {
+        self.pct(self.time_boost_25_50)
+    }
+
+    pub fn boost_50_75_pct(&self) -> f32 {
+        self.pct(self.time_boost_50_75)
+    }
+
+    pub fn boost_75_100_pct(&self) -> f32 {
+        self.pct(self.time_boost_75_100)
+    }
+
+    pub fn amount_obtained(&self) -> f32 {
+        self.amount_collected_big + self.amount_collected_small + self.amount_respawned
+    }
+
+    pub fn amount_used_by_vertical_band(&self) -> f32 {
+        self.amount_used_while_grounded + self.amount_used_while_airborne
+    }
+
+    pub(super) fn add_labeled_amount<I>(&mut self, labels: I, amount: f32)
+    where
+        I: IntoIterator<Item = StatLabel>,
+    {
+        if amount > 0.0 {
+            self.labeled_amounts.add(labels, amount);
+        }
+    }
+
+    pub(super) fn increment_labeled_count<I>(&mut self, labels: I)
+    where
+        I: IntoIterator<Item = StatLabel>,
+    {
+        self.labeled_counts.increment(labels);
+    }
+}
